+++ frontend/BE/src/main/java/com/a404/duckonback/service/ArtistFollowService.java
@@ -7,9 +7,9 @@ import java.util.Optional;
 
 public interface ArtistFollowService {
     ArtistFollow createArtistFollow(ArtistFollow artistFollow);
-    Optional<ArtistFollow> getArtistFollow(Long id, Integer artistId);
+    Optional<ArtistFollow> getArtistFollow(Long id, Long artistId);
     List<ArtistFollow> getFollowsByUser(Long id);
-    List<ArtistFollow> getFollowsByArtist(Integer artistId);
-    void deleteArtistFollow(Long id, Integer artistId);
-    boolean isFollowingArtist(Long id, Integer artistId);
+    List<ArtistFollow> getFollowsByArtist(Long artistId);
+    void deleteArtistFollow(Long id, Long artistId);
+    boolean isFollowingArtist(Long id, Long artistId);
 }
